% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{active}
\alias{active.default}
\alias{active<-}
\alias{active<-.default}
\title{Determine the context of subsequent manipulations}
\usage{
activate(.data, what)

active(x)

\method{active}{default}(x)

active(x) <- value

\method{active}{default}(x) <- value
}
\arguments{
\item{.data, x}{An object with named entities}

\item{what}{What should get activated? See methods for specifics.}

\item{value}{the name to be used as the active sub-object}
}
\value{
A object of class `.data`
}
\description{
Many objects are composed of multiple named entities and it necessary to specify which
sub-object is referenced during manipulations. The
`activate` verb does just that affects all subsequent manipulations
until a new sub-object is activated. `active` is a simple query function to get
the currently active context.
}
\examples{

# See dev tidygraph and dev ncdump for some real examples, but as seen
# we can activate a generic object by setting or getting attr(x, 'active').
my_object <- structure(list(x = 1, y = 25), active = "y")
active(my_object)
activate(my_object, "x")
}
